pyproject.toml
[project]
name = "ai-review"
version = "0.1.0"
description = "AI-powered code review tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "PyGithub>=2.1.1",
    "gitpython>=3.1.40",
    "click>=8.1.7",
    "openai>=1.3.0",  # Defaulting to OpenAI for now, can be changed
    "tenacity>=8.2.3",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
]

[project.scripts]
ai-review = "src.main:cli"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
